// src/models/ThemeColorEnum.ts
export enum ThemeColorEnum {
    REBALANCING_TOOL_LOGO = '#53FDCA',
    REBALANCING_TOOL_TEXT = '#121212',
    REBALANCING_TOOL_BACKGROUND = '#53FDCA',
    SUGGEST_APP_LOGO = '#FA8231',
    SUGGEST_APP_TEXT = '#121212',
    SUGGEST_APP_TEXT_BACKGROUND = '#FA8231',
    PURE_AIR_LOGO = '#FFFFFF',
    PURE_AIR_TEXT = '#FFFFFF',
    PURE_AIR_BACKGROUND = '#2994ff',
    DIEWELL_WEBSITE_LOGO = '#FFFFFF',
    DIEWELL_WEBSITE_TEXT = '#FFFFFF',
    DIEWELL_WEBSITE_BACKGROUND = '#0e243f',
    BROADY_PICTURES_LOGO = '#FFFFFF',
    BROADY_PICTURES_TEXT = '#FFFFFF',
    BROADY_PICTURES_BACKGROUND = '#383224',
}